requirements.txt
-e .
cloudpickle==1.6.0
cycler==0.11.0
dataclasses==0.8
gym==0.19.0
gym-classics==0.0.1
gym-notices==0.0.8
importlib-metadata==4.8.3
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyyaml==6.0
scipy==1.5.4
six==1.16.0
typing_extensions==4.1.1
zipp==3.6.0
